Agents: extract BaseAgentHostSessionsProvider to share local/remote logic (Written by Copilot)
Extracts the shared session/config/adapter/notification flow from
LocalAgentHostSessionsProvider and RemoteAgentHostSessionsProvider into
a new abstract BaseAgentHostSessionsProvider plus a single concrete
AgentHostSessionAdapter (with an options bag for variation points), both
under src/vs/sessions/contrib/agentHost/browser/.

The local provider drops from 1164 to 186 LOC. The remote provider
drops from 1457 to 395 LOC, retaining the connection-lifecycle surface
(setConnection / clearConnection / setAuthenticationPending /
setConnectionStatus / setOutputChannelId), the well-known agent-type
mapping, and the remote folder picker. Public API and behavior are
unchanged; both providers' existing test suites pass unmodified.
